# medibuddy-backend/MediBuddy-Backend-main/openapi.yaml
openapi: 3.0.3
info:
  title: MediBuddy API
  description: |
    REST API for MediBuddy — a medication management, symptom logging, and
    consultation-summary PWA designed for low-resource settings.

    ## Base URL
    All endpoints are prefixed with `/api`.

    ## Authentication
    Most endpoints require a **Bearer JWT** in the `Authorization` header.
    Obtain tokens via the OTP → verify flow (see Auth endpoints).

    ## Response Envelope
    Every response follows:
    ```json
    { "success": true, "data": { ... }, "message": "optional", "serverTime": "ISO-8601" }
    ```
    Errors:
    ```json
    { "success": false, "error": "msg", "code": "CODE", "details": [], "serverTime": "ISO-8601" }
    ```

    ## Idempotency
    Mutating endpoints accept an `Idempotency-Key` header (UUID v4).
    The server caches responses for 24 h; replaying the same key returns the cached result.

    ## Rate Limits
    | Preset | Window | Max |
    |--------|--------|-----|
    | General | 15 min | 100 |
    | Auth | 15 min | 20 |
    | OTP | 15 min | 5 |
    | Sync | 1 min | 30 |
    | Write | 1 min | 60 |

    Rate headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `Retry-After`.

  version: 1.0.0
  contact:
    name: MediBuddy Team
  license:
    name: UNLICENSED
servers:
  - url: http://localhost:5000/api
    description: Local development
  - url: https://medibuddy-backend-w9hf.onrender.com/api
    description: Production (Render)
  - url: http://localhost:4010
...
paths:
...
  /profiles:
    get:
      operationId: listProfiles
      tags: [Profiles]
      summary: List profiles owned by the user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of profiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                success: true
                data:
                  - _id: "aaa111"
                    ownerId: "usr001"
                    name: "John Doe"
                    dob: "1990-05-15"
                    relation: "self"
                    conditions: ["hypertension"]
                serverTime: "2026-02-10T12:05:00.000Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createProfile
      tags: [Profiles]
      summary: Create a new profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfile"
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /profiles/{profileId}:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: getProfile
      tags: [Profiles]
      summary: Get profile details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateProfile
      tags: [Profiles]
      summary: Update profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfile"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    delete:
      operationId: deleteProfile
      tags: [Profiles]
      summary: Soft-delete profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ MEDICATIONS ═══════
  /profiles/{profileId}/medications:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listMedications
      tags: [Medications]
      summary: List medications for a profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of medications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: createMedication
      tags: [Medications]
      summary: Create a medication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedication"
      responses:
        "201":
          description: Medication created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /profiles/{profileId}/reminders:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listReminders
      tags: [Reminders]
      summary: List reminders (24 h window)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of reminder events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /profiles/{profileId}/symptoms:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listSymptoms
      tags: [Symptoms]
      summary: List symptom logs with optional date range
      security:
        - bearerAuth: []
      parameters:
        - name: start
          in: query
          schema:
            type: string
            format: date-time
        - name: end
          in: query
          schema:
            type: string
            format: date-time
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated symptoms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: logSymptom
      tags: [Symptoms]
      summary: Log a symptom (supports idempotency)
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSymptom"
      responses:
        "201":
          description: Symptom logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ SUMMARIES ═══════
  /profiles/{profileId}/summaries:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listSummaries
      tags: [Summaries]
      summary: List consultation summaries
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of summaries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: generateSummary
      tags: [Summaries]
      summary: Generate a consultation summary PDF
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [startDate, endDate]
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
            example:
              startDate: "2026-01-01"
              endDate: "2026-02-01"
      responses:
        "201":
          description: Summary generation started / completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /profiles/{profileId}/caregivers:
    parameters:
      - $ref: "#/components/parameters/profileId"
    get:
      operationId: listCaregivers
      tags: [Caregivers]
      summary: List caregivers for a profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of caregivers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
    post:
      operationId: addCaregiver
      tags: [Caregivers]
      summary: Add a caregiver
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaregiver"
      responses:
        "201":
          description: Caregiver added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /messages:
    post:
      operationId: sendMessage
      tags: [Messages]
      summary: Send a message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessage"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /messages/system:
    post:
      operationId: sendSystemMessage
      tags: [Messages]
      summary: Send a system message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [profileId, content]
              properties:
                profileId:
                  type: string
                content:
                  type: string
                type:
                  type: string
                  default: "system"
      responses:
        "201":
          description: System message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /messages/conversations:
    get:
      operationId: listConversations
      tags: [Messages]
      summary: List conversations for the user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /messages/unread:
    get:
      operationId: getUnreadCount
      tags: [Messages]
      summary: Get unread message count
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
              example:
                success: true
                data: { count: 3 }
                serverTime: "2026-02-10T12:15:00.000Z"
  /messages/conversation/{conversationId}:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getConversation
      tags: [Messages]
      summary: Get messages in a conversation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Array of messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /messages/conversation/{conversationId}/read:
    parameters:
      - name: conversationId
        in: path
        required: true
        schema:
          type: string
    patch:
      operationId: markConversationRead
      tags: [Messages]
      summary: Mark conversation as read
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /messages/{messageId}:
    parameters:
      - name: messageId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteMessage
      tags: [Messages]
      summary: Delete a message
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"

  # ═══════ PUSH ═══════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Obtain via `POST /auth/verify`. Short-lived access token (15 min).
        Use `POST /auth/refresh` with the refresh token to get a new pair.
  parameters:
    profileId:
      name: profileId
      in: path
      required: true
      schema:
        type: string
      description: MongoDB ObjectId of the profile
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: UUID v4 — server caches responses for 24 h
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              - path: "phone"
                message: "Invalid phone number format"
            serverTime: "2026-02-10T12:00:00.000Z"
    Unauthorized:
      description: Authentication required or token expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "No token provided"
            code: "UNAUTHORIZED"
            serverTime: "2026-02-10T12:00:00.000Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
            serverTime: "2026-02-10T12:00:00.000Z"
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, serverTime]
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response payload (varies per endpoint)
        message:
          type: string
        serverTime:
          type: string
          format: date-time
    ErrorEnvelope:
      type: object
      required: [success, error, code, serverTime]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
        serverTime:
          type: string
          format: date-time
...
    CreateProfile:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Amina Ibrahim"
        dob:
          type: string
          format: date
        relation:
          type: string
          enum: [self, family, other]
        conditions:
          type: array
          items:
            type: string
    UpdateProfile:
      type: object
      properties:
        name:
          type: string
        dob:
          type: string
          format: date
        conditions:
          type: array
          items:
            type: string
    CreateMedication:
      type: object
      required: [name, times]
      properties:
        name:
          type: string
          example: "Amlodipine"
        dosage:
          type: string
          example: "5mg"
        times:
          type: array
          items:
            type: string
          example: ["08:00", "20:00"]
        timezone:
          type: string
        channels:
          type: array
          items:
            type: string
            enum: [push, sms]
          default: [push]
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        clientId:
          type: string
        clientTs:
          type: string
          format: date-time
...
    CreateSymptom:
      type: object
      required: [symptom, severity, clientTs]
      properties:
        symptom:
          type: string
          example: "headache"
        severity:
          type: integer
          minimum: 1
          maximum: 5
          example: 3
        duration:
          type: string
          example: "2 hours"
        notes:
          type: string
        voiceNoteUrl:
          type: string
        clientId:
          type: string
        clientTs:
          type: string
          format: date-time
    CreateCaregiver:
      type: object
      required: [name, phone]
      properties:
        name:
          type: string
          example: "Dr. Adebayo"
        phone:
          type: string
          example: "+2348098765432"
        relation:
          type: string
          example: "doctor"
        alertOnMissedDose:
          type: boolean
          default: true
        alertThresholdMinutes:
          type: integer
          default: 30
...
    CreateMessage:
      type: object
      required: [profileId, content]
      properties:
        profileId:
          type: string
        content:
          type: string
        type:
          type: string
          default: "text"